Riiiiight int sizing and the sign bit

+++ lab1.c
@@ -91,8 +91,8 @@ void Control(void){
        int len = sizeof(BufferLastEvent) / sizeof(BufferLastEvent[0]);
        int x = len;
       for(x; x >= 0 && Flags > 0; x--) {
-        int ex = exp2(x);
-        printf("\n\tCHECK: %d -> %d", ex);
+        long unsigned int ex = exp2(x);
+        printf("\n\tCHECK: %d -> %u", ex);
         printf("\n\tFLAGS: %d", Flags);
 
         // if the current flag is not set, don't handle an event
